+++ SmallStep/Core/SmallStep.h
@@ -31,9 +31,6 @@
 // Generic desktop main menu
 #import "SSMainMenu.h"
 
-// GNUstep/AppKit compatibility (NSView setTag:/tag/viewWithTag:)
-#import "NSView+SSTag.h"
-
 // Version
 FOUNDATION_EXPORT double SmallStepVersionNumber;
 FOUNDATION_EXPORT const unsigned char SmallStepVersionString[];